Handle incorrect network CIDRs more gracefully
Currently if you attempt to create an infoblox_ip resource using a
CIDR that is not a managed network in infoblox, terraform will crash
with an unexpected EOF error. This is because the call to
`client.Network().Find()(q, nil)` returns an empty array instead of an
error. Later the line `ou, err := client.NetworkObject(out[0]["_ref"]
.(string)).NextAvailableIP(1, nil)` tries to access a non-existent
elemtn in the array.